v3.1 sdk

+++ esp32/build/build-idf.sh
@@ -7,7 +7,10 @@
 esp_idf_branch=${1:-v3.1}
 #espruino_branch=${2:-master}
 espruino_branch=${2:-ESP32-V3.1}
-echo using esp-idf branch $esp_idf_branch, espruino branch $espruino_branch
+checkout_mode=${3:-none}
+# Fresh cleans down to get new release
+#checkout_mode=$(3:-fresh}
+echo using esp-idf branch $esp_idf_branch, espruino branch $espruino_branch, checkout mode $checkout_mode
 
 if [ ! -d "Espruino" ]; then
 git clone https://github.com/espruino/Espruino.git
@@ -26,24 +29,31 @@ cd ..
 # This will need to be tied to a release
 #git submodule update --init 
 
-
 #cd esp-idf
 #git fetch
 #git checkout $esp_idf_branch
-rm -r esp-idf
-git clone -b $esp_idf_branch --recursive https://github.com/espressif/esp-idf.git esp-idf
-git submodule update --init --recursive
-cd ..
+if [ "$checkout_mode" = "fresh" ]
+then
+        echo "refreshing to new sdk"
+	rm -r esp-idf
+	git clone -b $esp_idf_branch --recursive https://github.com/espressif/esp-idf.git esp-idf
+	git submodule update --init --recursive
+fi
+
+echo `pwd`
+
 # adjust paths to this folder versions
 export ESP_IDF_PATH=`pwd`/esp-idf
 export IDF_PATH=`pwd`/esp-idf
 export ESP_APP_TEMPLATE_PATH=`pwd`
 cd app
 make clean
-make
-#make -j 5
+#make
+make -j 5
 # This is not the firmware - get rid of it!
-rm build/espruino-esp32.bin
+rm build/espruino-esp32.elf
+echo `pwd`
+cd ..
 make app.tgz
 cd ../Espruino
